requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_py-xgboost-mutex=2.0=cpu_0
absl-py=1.0.0=pypi_0
adjusttext=0.7.3.1=py_1
anyio=2.2.0=py38haa95532_2
appdirs=1.4.4=pyhd3eb1b0_0
argon2-cffi=20.1.0=py38h2bbff1b_1
async_generator=1.10=pyhd3eb1b0_0
atomicwrites=1.4.0=pypi_0
attrs=21.2.0=pyhd3eb1b0_0
audioread=2.1.9=py38haa244fe_0
augmentor=0.2.9=pypi_0
autofeat=2.0.10=pypi_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
bidict=0.21.3=pypi_0
biosppy=0.7.3=pypi_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2021.10.8=h5b45459_0
cachetools=5.0.0=pypi_0
certifi=2021.10.8=py38haa244fe_2
cffi=1.15.0=py38h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cloudpickle=2.0.0=pyhd8ed1ab_0
colorama=0.4.4=pyhd3eb1b0_0
cryptography=35.0.0=py38h71e12ea_0
cudatoolkit=11.3.1=h59b6b97_2
cvxopt=1.2.7=pypi_0
cvxpy=1.1.17=pypi_0
cvxpy-base=1.1.18=py38h5d928e2_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.5.1=py38hd77b12b_0
decorator=5.1.0=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
deprecation=2.1.0=pyhd3eb1b0_0
docopt=0.6.2=pypi_0
ecos=2.0.7.post1=pypi_0
entrypoints=0.3=py38_0
et_xmlfile=1.1.0=py38haa95532_0
fancyimpute=0.7.0=pypi_0
ffmpeg=4.3.1=ha925a31_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
future=0.18.2=pypi_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
google-auth=2.6.6=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.44.0=pypi_0
gym=0.21.0=py38h595d716_2
h5py=3.6.0=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=3.3=pyhd3eb1b0_0
imitation=0.2.0=dev_0
importlib-metadata=4.8.1=py38haa95532_0
importlib_metadata=4.8.1=hd3eb1b0_0
iniconfig=1.1.1=pypi_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.4.1=py38haa95532_1
ipython=7.29.0=py38hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jdcal=1.4.1=pyhd3eb1b0_0
jedi=0.18.0=py38haa95532_1
jinja2=3.0.2=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h2bbff1b_0
json5=0.9.6=pyhd3eb1b0_0
jsonpickle=1.5.2=pypi_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter-packaging=0.10.4=pyhd3eb1b0_0
jupyter_client=7.0.6=pyhd3eb1b0_0
jupyter_core=4.9.1=py38haa95532_0
jupyter_server=1.4.1=py38haa95532_0
jupyterlab=3.0.14=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.8.2=pyhd3eb1b0_0
kiwisolver=1.3.1=py38hd77b12b_0
knnimpute=0.1.0=pypi_0
libblas=3.9.0=1_h8933c1f_netlib
libcblas=3.9.0=5_hd5c7e75_netlib
libflac=1.3.3=h0e60522_1
liblapack=3.9.0=5_hd5c7e75_netlib
libogg=1.3.5=h2bbff1b_1
libopus=1.3.1=h8ffe710_1
libpng=1.6.37=h2a8f88b_0
librosa=0.8.0=pyh9f0ad1d_0
libsndfile=1.0.31=h0e60522_1
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libvorbis=1.3.7=he774522_0
libwebp=1.2.0=h2bbff1b_0
libxgboost=1.3.3=hd77b12b_0
lightgbm=3.2.1=py38hd77b12b_0
llvmlite=0.34.0=py38h1a82afc_4
lz4-c=1.9.3=h2bbff1b_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.6=pypi_0
markupsafe=2.0.1=py38h2bbff1b_0
matplotlib=3.5.0=py38haa95532_0
matplotlib-base=3.5.0=py38h6214cd6_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mido=1.2.10=pypi_0
mistune=0.8.4=py38he774522_1000
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
more-itertools=8.10.0=pypi_0
mpmath=1.2.1=pypi_0
msys2-conda-epoch=20160418=1
munch=2.5.0=pypi_0
munkres=1.1.4=py_0
music21=7.1.0=pypi_0
muspy=0.4.0=pypi_0
nbclassic=0.2.6=pyhd3eb1b0_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py38haa95532_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
nose=1.3.7=pypi_0
notebook=6.4.6=py38haa95532_0
numba=0.51.2=py38h251f6bf_0
numpy=1.21.4=pypi_0
numpy-base=1.21.2=py38h0829f74_0
oauthlib=3.2.0=pypi_0
olefile=0.46=pyhd3eb1b0_0
opencv-python=4.5.4.58=pypi_0
openpyxl=3.0.7=pyhd3eb1b0_0
openssl=1.1.1n=h8ffe710_0
osqp=0.6.2.post0=py38h5d928e2_3
packaging=21.3=pyhd3eb1b0_0
pandas=1.2.4=py38hd77b12b_0
pandocfilters=1.4.3=py38haa95532_1
parso=0.8.2=pyhd3eb1b0_0
pickle-mixin=1.0.2=pypi_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py38hd45dc43_0
pint=0.18=pypi_0
pip=21.2.2=py38haa95532_0
pluggy=1.0.0=pypi_0
pooch=1.4.0=pyhd3eb1b0_0
pretty-midi=0.2.9=pypi_0
prometheus_client=0.12.0=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.20.1=pypi_0
py=1.11.0=pypi_0
py-cpuinfo=8.0.0=pypi_0
py-xgboost=1.3.3=py38haa95532_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.8.2=py38h2bbff1b_0
pyglet=1.5.21=pypi_0
pygments=2.10.0=pyhd3eb1b0_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pypianoroll=1.0.4=pypi_0
pyqt=5.9.2=py38ha925a31_4
pyrsistent=0.18.0=py38h196d8e1_0
pysocks=1.7.1=py38haa95532_0
pysoundfile=0.10.3.post1=pyhd3deb0d_0
pytest=6.2.5=pypi_0
python=3.8.8=hdbf39b2_5
python-dateutil=2.8.2=pyhd3eb1b0_0
python-slugify=5.0.2=pypi_0
python_abi=3.8=1_cp38
pytorch=1.10.0=py3.8_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pywavelets=1.1.1=py38he774522_2
pywin32=228=py38hbaba5e8_1
pywinpty=0.5.7=py38_0
pyyaml=6.0=pypi_0
pyzmq=22.2.1=py38hd77b12b_1
qdldl=0.1.5.post0=pypi_0
qdldl-python=0.1.5=py38h5d928e2_2
qt=5.9.7=vc14h73c81de_0
requests=2.26.0=pyhd3eb1b0_0
requests-oauthlib=1.3.1=pypi_0
resampy=0.2.2=py_0
rsa=4.8=pypi_0
sacred=0.8.2=pypi_0
scikit-learn=1.0.1=py38hf11a4ad_0
scipy=1.8.0=py38ha1292f7_1
scs=2.1.4=pypi_0
seaborn=0.11.2=pyhd3eb1b0_0
seals=0.1.2=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py38haa95532_0
shortuuid=1.0.8=pypi_0
sip=4.19.13=py38ha925a31_0
six=1.16.0=pyhd3eb1b0_0
smmap=5.0.0=pypi_0
sniffio=1.2.0=py38haa95532_1
sqlite=3.36.0=h2bbff1b_0
stable-baselines3=1.4.1a0=pypi_0
sympy=1.9=pypi_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
terminado=0.9.4=py38haa95532_0
testpath=0.5.0=pyhd3eb1b0_0
text-unidecode=1.3=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h2bbff1b_0
toml=0.10.2=pypi_0
tomlkit=0.7.2=py38haa95532_1
torchaudio=0.10.0=py38_cu113
torchvision=0.11.1=py38_cu113
tornado=6.1=py38h2bbff1b_0
tqdm=4.62.3=pypi_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.7=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
webcolors=1.11.1=pypi_0
webencodings=0.5.1=py38_1
werkzeug=2.1.1=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38haa95532_2
winpty=0.4.3=4
wrapt=1.14.0=pypi_0
xgboost=1.3.3=py38haa95532_0
xlrd=2.0.1=pyhd3eb1b0_0
xz=5.2.5=h62dcd97_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
